+++ src/video/DummyRenderer.hh
@@ -41,7 +41,7 @@ public:
 	void updateSpritesEnabled(bool enabled, const EmuTime& time);
 	void updateVRAM(unsigned offset, const EmuTime& time);
 	void updateWindow(bool enabled, const EmuTime& time);
-	virtual float getFrameRate() const;
+	virtual double getFrameRate() const;
 
 	// Layer interface:
 	virtual void paint();
